+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-# 2.5.4
-Fix https://github.com/MCTeamPotato/Kasualix-Issue-Tracker/issues/207
+# 2.5.5
+- Implement https://github.com/MCTeamPotato/Kasualix-Issue-Tracker/issues/207#issuecomment-3699981669 for Infernal Forge and Gauntlet Of Guard
+- Implement https://github.com/MCTeamPotato/Opotato/issues/8#issuecomment-3699995182 for Brutal Bosses bugfix